Themes pages into Bootstrap / C5 styles.
Have some fun here! Just updated the styles to Bootstrap-friendly in some cases...played it a little fast and loose with what was there for the 'Inspect' page. Extra CSS tweaks above and beyond Bootstrap are, again, inline...will need to be added to .less > .css in due course, assmuming approved. One bug -- whilst i've dropped with width of the pages (span12 & span14), the success message on install of theme, etc... is full width. Looks a little odd is all.