Theme Customization Dashboard page rebuild + styling
Quite a few changes here...probably got carried away. Updating styling -- nice and simple and fairly Bootstrap-y. Few visual/layout tweaks to try and make the most of the more limited preview space due to the new look of C55 (fixed width dashboard pages). Stacked the style editors on top of the preview, as width is where we're most challenged here. Also made collapsible styles by hooking into the togglePane JS functionality. Bear in mind, I haven't touched the styling of the opening customization panels themselves yet (probably not worth breaking too much sweat over it, but could perhaps change the buttons to match, if nothing else)...just styling of the layout. Tried to make as flexible as possible, without resorting to fixed-wdith (few ugly mods and hacks to make IE7 play ball as much as poss).
What you will notice is a nifty new 'Fullscreen Preview' button in the footer. This is what i've spent ALOT of time fiddling with -- figured it would be nice to be able to get a big view of the theme during customisation, just to get a better sense of what the tweaks are doing to the design... copied the openPreview JS from the Themes landing page, and modified (JS currently inline...will tidy and fix assuming you don't think this is a total waste of time!). Have a play and let me know what you think. For a simple life, we just lose the fullscreen business and use the basic styling i've done OR make the preview pane a little taller in situ and break the styles panel out into a modal (as per fullscreen) on page load -- code is commented out for this in the document.ready JS at the mo.


Former-commit-id: 3f90a3ac43ca53e3ef951a2d4fbbec1821247dba