Tidy up Logs page to bring more inline with survey + forms reports.
Juggled this around a bit to hopefully make more consistent and make more sense. Search fields you might want to keep in, but i've moved them to ccm-options-pane, leveraged the ccm-options-pane-permanent-search styling you've got in for sitemap search, and only show them (and logs table) when logs exist. Kept the central alignment on the 'clear log' button, but added 'large' class. Fixed the pagination that was completely failing in bootstrap styles thanks to C5 pagination helper spitting out spans instead of li (semantically, lists ul>li are def better than spans, so probably worth making the shift). To this end, i've added a variable to the paginaion helper that will allow to pass 'li' to the getPages() call to get it to use them as wrappers. Changed nothing else and should be backwards compatible -- added a little inline CSS styling to cope with the other stuff C5 spits out there (eg. no link on the currently active page) to make work with Bootstrap. Generate a load of logs and let me know what you think!