Tweak height of custom template edit dialog
Former height of modal dialog (110 pixels) was insufficient when applying a custom template to a global area -- the notice that the editor is changing the template on a global area block pushed the select field down, requiring ugly scroll. Increased height to accommodate. Leaves some whitespace when NOT editing a global area, but not the end of the world...probably better this way than risking editors not realising a scroll in the modal / a drag-resize required.


Former-commit-id: 14d898fcbd9000f56dab4b4e6d9bc1a2b78b9ed1