File Sets List - spacing tweak to avoid 'add file set' button overlap
Forgive the hasty use of inline CSS here, but it does the trick! When the file sets page has file sets to list, the 'add file set' button on the right hand side overlaps the :hover on the div/row of first listed set. Just moved the button -5px (which actually helps to line up centrally with the 'no file sets found' or the paging message, thanks to the line-height) and added an extra 5px to the padding-bottom of the ccm-paging-top div.