Colorpicker buttons - Bootstrap styling
Submitted a pull for this an age ago when we were still working with the closed Alpha, but think it might have got lost in the move to public beta. Had a bit of time now to pull the CSS out again, check it's compatible with the latest version and stick back up here. Taken pains to make sure these should provide the styling to any implementation of the .colorpicker class and the core's JS colorpicker functionality, whether or not it has access to the .ccm-ui class. Haven't done the compile from LESS as i'm liable to do it wrong, but has compiled fine on my test server and a couple of screengrabs are below. Take it or leave it, just makes the tiny buttons a little snazzier!