Editing toolbar link nesting fix
Took me a while to squirrel this one out! Noticed that buttons, search field alignment and links were thrown off in the toolbar when an editor without Dashboard access/permissions put a page into edit mode. Turns out that the link on the edit/editing button on the left hand side wasn't being correctly closed in all cases, causing the edit mode link to encapsulate the search field on the right-hand side of the toolbar when activated ('editing'). All fixed now!