don't implicitly trust that a selectedSearchField $item is an attribute key that still exists
I had to hack this in to get the site wide search to work after I had set up a search using the blog add-on's select attributes, I uninstalled the package and the search stuck around and broke search entirely. This might be more of a cache invalidation issue but this fixed the issue here.


Former-commit-id: c881871714683e563c5af1135046a0ffec22afd4