checkbox for form notification from address
Allow form submission notifications to be sent from a specific email field in
the form.

An associative array of options is saved to the btFormQuestions table to
save this preference. In the future if more options are required for the email
(or other field types) this methodology could be used. Otherwise a new
boolean column could be added to the record

Former-commit-id: c3dc11687e0355bc69aef27d2ab57c7fbea3351b