Fix issue with closeAll running in reverse:
New dialogs are appended to the dom, meaning the higher the dialog level, the farther down the dom it is.
Running `.get()` grabs the JavaScript HTML Elemnts, prime for running `.reverse()` to fully reverse the array, then enclosing that in `$()` converts it back to a manipulable jQuery object.

This becomes useful when you have 2 dialogs stacked wherein the lower dialog manages the results of the higher dialog. (Example case is eCommerce, issue #16)
Top Dialog manipulates value in span, bottom dialog grabs that value and verifies, then places it in an input to be saved. Running it in the current environment gives opposite results.


Former-commit-id: 7f19a08bd8ab39b7c4d486ef18539f99c4c1f190