Dashboard menu/toolbar fix to respect News page permissions
Fix and tidy of permission checks determining what's shown in the footer of the JS dashboard dropdown menu. No matter the permissions, non-admins without viewing access were still being shown the /news page link in the footer of the dropdown...now permissions are checked first. Also tidied the permissions check on all 3 links in there to allow us to skip dropping in a UL if no links required.


Former-commit-id: 025472c8e709faf251b0696ce0d36262fd0af6a6