Correct ordering of log list.
Fix issue where multiple log entries with the same timestamp can display in a non-chronological order. By imposing ordering by logID (176, 178), the log listing is now always in teh order in which it was made.

The 'order by timestamp' part is no longer necessary, but I have left it in as it is not performance critical and provides a bit of redundant defensive code.

Former-commit-id: 1a2b7ac20735fdd0000ecd713f86a7746b039e17