To aid backtracking of original images from cached images / thumbnails.
Re-worked pull request 588 that I made on the master branch into
the 553 devel branch.
https://github.com/concrete5/concrete5/pull/588

Scripts in the browser will now be able to backtrack a thumbnail to the
original image.

Defaults to existing cache file naming. I am not sure if switching
behaviour on IMAGE_HELPER_APPEND_ORIG_FID is necessary. A better
solution may be to forget about the existing cache file naming and just
get on with appending fID when it is available.

Anyway, the change is coded so only if IMAGE_HELPER_APPEND_ORIG_FID is
defined and not false, then appends "_f".$fID to the cache file name.
When enabled, scripts in the browser will now be able to backtrack a
thumbnail to the original image.


Former-commit-id: 6e5a4ba7798325b30c6266e7fa39795cadcaff2b