Select Attribute - sanitization/duplication fix
This removes the output specialchars SelectOption sanitization by boolean in the Select Attribute controller.

Why? Well...a bug's been annoying me (via my clients) for quite a while. Basically, if you create a Select type attribute and allow user addition and multiple selection, if they create any with an ampersand (&), the current sanitization (specialchars) that's in place on the getSelectAttributeOptionValue() causes the pre-existing option check to fail (& != &amp;) and multiple options to be allowed through (as C5 does not actually go on to sanitize the user input, and the un-sanitized value is stored regardless). This is MOST annoying in Composer as, with the regular autosaves, a new version of this attribute option is saved each time, leading to some scarily-long attribute option lists, full of duplicated '&' containing entries.


Former-commit-id: 86c22fca3312ed2313d8a482ee953cde707e615c