Add on_page_body_index event
This event is passed the current collection which is being indexed. Any
content which is returned by the hook is appended to the content for the
search index.

This was necessary for a project where there was some content
programatically added to pages which needed to be included in the search
index.


Former-commit-id: 8dc40cccaf3e59e035a385605ad0f4a5c1ee9947