display all pages by default in pagelist block
There's an internal limit of 20 by default. If we don't call
setItemsPerPage, we only get 20 pages but this isn't clear in the
block dialog. I simply removed the check and call setItemsPerPage(0)
if no value has been entered.
We could also display the number 20 when adding a new block, either
way is okay I think.

Former-commit-id: e0462fa21f585c39a963cea63ce8f20454147215