Fix Form Results Page Width Issue
This changes the form results page
by wrapping the form results table
in a div with a set width and
letting it scroll. The scroll bar
is set to always show in webkit
and a message appears notifying the
user when there is content offscreen
that can be accessed by scrolling.

This solution keeps the date sortable
row intact which we would lose if we
went with rows headings on the left
side.