Some Select Attribute Type UI/UX TLC!
Smartened up the default front-end presentation of those Select type attributes to bring them more in line with the rest of the UI and the Bootstrap styles. Always chaffed that they didn't get much love in the big 5.5 push and just building some custom Single Page interface for a client...figured i'd have a shot a smartening the core up a bit instead of just overriding for my use case. Nothing too flash and no structural changes whatsoever...just a few classes and some .ccm-ui namespaced CSS styles here and there.

Testing and rendering fine in latest Chrome, Firefox and IE down to IE7 (where the rest of the Bootstrap stuff starts to look a little ropey anyway).

Only one JS change - I added in a length check on press of the 'Add' button. Easier trying to stop users mistakenly submitting those ugly empty values to the DB (especially annoying when they tend to get duplicated in there, and leaving lots of very weird looking empty options when the attribute is listed/rendered) on the client-side. Certainly when a check already existed on Enter key-press but not, for some reason, on click of the actual 'Add' button. I've used the exact same approach to the string existence check as used in the existing, core key-press validation...works fine for me across browser and shouldn't cause any headaches for a quick, simple merge!