remove "block wrapper tokens" functionality
Hacky thing I added back in 5.5 or 5.6 -- was a mistake, I shouldn't have put it in there (sorry!). I don't expect too many people are utilizing the feature since it was never announced or documented. Seems like un-necessary complexity in the core now, and I figure 5.7 is as good a time as any to make a clean break (since there will already be backwards-incompatibilities it sounds like).