Fix adding timestamp in ccm_addHeaderItem
When adding a timestamp to the resources (.css/.js files) loaded by
ccm_addHeaderItem, add a timestamp to avoid cache.
BTW, if the URL of the resource already contains a GET parameters
(eg 'file.js?a=X'), the timestamp is appended withoud adding a new
parameter (eg 'file.js?a=X1393878504000).
That will mess up the existing GET parameter.

Let's fix this by adding a new GET parameter when adding the timestamp
even when we already have other GET parameters.