Throw an exception if importing an unknown block
When importing CIF data that contains an unknown block type, the execution
halts with an error message that does not help understanding what happened
("getController() method called on a non object").
This may happen for instance we don't have installed a package that
implements a specific block type.

It's better if we throw an exception that explains better what went wrong.


Former-commit-id: 28c85ab01aac4770b67d04f06147578885963973