Missing typography.css throws exception
In theme `twitter_bootstrap` exception is thrown on first generation of
`getStyleSheet()`

The theme header calls `$view->getStyleSheet('typography.css')` but
typography.css does not exist so error is thrown. It is very likely
a better solution will be added upstream to check for the file, but did
not want to overengineer it.

- Add empty typography.css file