Fix getIntelligentSearchMenu cache
getIntelligentSearchMenu checks if a value called "dashboardMenus" exists
in cache, but if it does not exists it caches the value
"dashboardMenus/locale".
This causes getIntelligentSearchMenu to not use cache.

Let's fix this, by using an array with key "dashboardMenus": so, clearing
the cache simply requires to remove the "dashboardMenus" array from the
cache (instead of removing all the possible "dashboardMenus/locale" keys).