Fix labels in cache settings page
The "for" attribute of <label> elements must contain the id of the element
for which they are a label.
Currently they are referring to the names attributes, so they don't work.

Let's remove the "for" attribute, since the elements for which they are
labels are contained within the <label> elements themself.