Fix the page selector for good
I took a bit of time to look over the code that I had, and try to
re-concieve the mangled ideas that I had before, and ended up realizing
my mistake. I was unbinding the `ConcreteSitemap` event on first hit
which would break the second `ConcreteSitemap` tab. What I've done now
is namespaced all of those bindings, and left them open
until we get a concrete result (which would be triggered by the dialog
closing.)


Former-commit-id: eac319cbb51c1557a3005377197c877f9a88f730