Remove unused $valueList parameter from Stylesheet
Remove unused `$valueList` paramenter from
`Concrete\Core\StyleCustomizer\Stylesheet`

`$valueList` appears to have become a class property now as opposed to
being passed in as a parameter. The parameter does not appear to be used
in the method and a scan for other `getCSS()` calls shows them as also
not using the parameter.


Former-commit-id: fd8ed78e917cf55514800b225a6fd7b1c6c87af8