Remove unused $valueList parameter from Stylesheet
Remove unused `$valueList` paramenter from
`Concrete\Core\StyleCustomizer\Stylesheet`

`$valueList` appears to have become a class property now as opposed to
being passed in as a parameter. The parameter does not appear to be used
in the method and a scan for other `getCSS()` calls shows them as also
not using the parameter.


Former-commit-id: f7064633b90f0331d280388b8a6ac086b510e6de