Bugfix: Use default page template if available
When adding a page (probably only programmatically), the default page template was not respected. Adding a page of type 'X' would result in adding a page of type X with a blank/no template, which would in turn display as the default template. I added one line of code to resolve that, so it now uses the page type's default template if we know it. The template can still be overridden by manually providing the `$template` parameter.

Former-commit-id: 944fa2daf3abee9a2fd0ef44cf85c493990e62cd