Use default template on page add when none provided
This is a more comprehensive version of the previous commit. It adds code for when no `$template` parameter is provided to the `add()` function. When that parameter is omitted, and when the user did provide a page type, then it will now use the page type's default template. 

This is an improvement over the previous version because it fixes the problem rather than the symptom, and as such fixes dependent behavior, such as that it will now copy the page type's default blocks as well (which it didn't in the previous fix).