Always use custom view if provided
In the (new?) 5.7 version of BlockView, you can still provide a custom filename to $bv->render(), but it isn't always honored. It is only honored if the BlockView also has a block attached to it. If a BlockView doesn't have a block attached to it, then the custom template you requested will not be rendered, but it will render to the default "view.php" instead. You can circumvent this behavior by providing the custom template in $controller->blockViewRenderOverride, but that's just a workaround, not a solution. If the BlockView allows you to pass in a template in the 'render' method, then it should try to honor it. I added some code that does just that.