Always use custom view if provided
The previous commit was only half a solution to the problem: you still couldn't pass in custom templates names that referred to a directory. It would break if you requested to render "blocks/mine/templates/alpha/view.php" by doing $bv->render("templates/alpha"). That was possible in Concrete < 5.7, and the above adjustments make it possible in 5.7 as well. It's much cleaner and closer to the logic in 5.6.3.1 as a bonus.