File manager should always return a value when form is submitted
If you would clear a file from the file manager popup, it would not get saved. If you would reload the page and edit the block, the file would still be there. The reason for that is that the file manager would return no value when no file is present. I think it's better to change that, so you can add the file manager to a block and have the 'clear' functionality work out of the box.