Addition of attribute handles as link titles
When developing with attributes a common task is to look up for
existing attributes their handles for use in code.

This change adds to the lists of attributes within the dashboard, the
attribute handle as a title attribute.

This means that instead of a developer having to repeatedly view the
edit page an attribute just to see the handle, they can mouse-over the
links to see this.


Former-commit-id: d10cb3acdad3c44b5567b27db7b3a2cc917acf84