Suggested tweaks to toolbar button labels
Adjustments to toolbar accessibilities titles
This is a suggestion for more meaningful labels for the main toolbar
buttons, more inline with their tooltips and purpose.

‘Add Page’ change to ‘Pages’  - This panel also has a very handy
sitemap in it, so it’s not just about adding pages, it’s about
navigating the pages of your site.

‘Settings’ change to ‘Page Settings’ - this button is referring
specifically to the current page, whereas the word Settings seems too
broad and could be interpreted to mean site settings.

‘Add’ to ‘Add Content’ - this button is always about adding blocks (or
stacks) to a page, but the word add by itself could imply other things.
Adding the word Content gives more context.

I’m sorry to be pedantic about these labels, but since these labels are
about improving accessibility they should be as meaningful as possible.


Former-commit-id: da05e7f89e413df411fa21a4c63efe14c417febf