Remove empty h2-tag, add placeholder alt tag
If a user is not placing a title for each slide, the h2 tag of the title is empty. This causes a warning with w3c-validation.
If a user is not placing a title for each slide, there is no alt-tag inside img-tag. This causes validation-error with w3c-validation.

In my fix the h2-tag is not shown, when the title of the image is empty. Same for the alt-tag.