Scroll to bottom of page when adding a new
When the page attributes panel is open and an attribute is added to the
active list, there is no visual feedback that something has been added.
The newly added attribute is added to the bottom of the page, but this
is normally beyond the viewport, requiring a scroll to see it.

This line automatically scrolls the viewport to the bottom, showing the
newly added attributes.