allow a page to be passed to getServiceLink
we're wanting to have a 'share this' on each instance of a page
in a page list template and thought it would be nice to use the code
that is in core already, but this would only share the page the user
is currently viewing. This PR would allow for a custom page to
optionally be passed in to override this behaviour.


Former-commit-id: 4e3b0998fc2a66e93710b04cf371dd373b3321c6