Use PermissionKeyCategory for pkgID in load()
Query takes in the `pkgID` from `PermissionKeys` which causes an
Exception as a permission key uses the `PermissionKeyCategory` if there
is not a custom permission key class set.

For example a new permission Key that is set to the `admin` permission
key category uses the `AdminKey` class by default. The current code
looks for `AdminKey` class in the package path if the permission key(not
the permission key category) belongs to a package.

- Use `PermissionKeyCategory.pkgID` for `pkgID` instead of
`PermissionKey.pkgID` in `PermissionKey::load()` and
`PermissionKey::loadAll()` queries.


Former-commit-id: 74c912bd05ee0f34fe31580048d0dce37da351bb