Add plural cases info
Let's add the name of the plural cases.
Until now we have that, for instance, Russian has 3 plural cases, identified by 0, 1 and 2.
That's not a great help when translators have to translate plurals.
It's much better to identify plurals by 'one', 'few' and 'others' (the CLDR identifiers)
and have some sample numeric range to explain when we have 'one', 'few' and 'others'.