Set of unique pagination page param for FileLists
If a file set is paginated (e.g. in a block like List Files From Set,
or in somewhere like the File Manager), it uses the ccm_paging_p as the
URL parameter to control which ‘page’ of results to display.

This parameter is the same used for the pagination of Page List blocks,
which means that if both are on the same page the blocks effectively
have linked pagination (incorrect behaviour).

By overriding in the FileList class the default parameter with
something unique, this problem is removed.

(there may be other classes that extend ItemList where this would be
worthwhile too)