A Block with CustomTemplate with No CSS CustomStyle should not print out ccm-custom-style-container div
I had an exact same case as
https://www.concrete5.org/community/forums/customizing_c5/is-it-possible-to-remove-the-div-which-surrounds-all-blocks-ccm-/

## What

We usually have original design with original HTML coding.
We want the blocks to print out exactly the way how we wrote in the HTML coding.

But we have multiple coding for a particular block (e.g., Page List blocks). So we make multiple custom templates.

But we don't want the extra ccm-custom-style-container div tag to be printed out because it will break the CSS or JS.

## About my PR

I know that my PR is premature.
I should be looking where it saves the custom template.

But I was only able to track the output side and implement for my client work.

Since I'm running out of my time, I would first throw this ugly PR, so that we can start a discussion.