getRelativePathFromID: 'false' instead of error
If you call 'getRelativePathFromID' for a non-existing fID, the file would evaluate to NULL, and you can't get a relative path from NULL -> this would result in a fatal error. I've added code to return 'false' in such cases instead. This kind of behavior could happen if you've transferred the database but not the files itself. The previous behavior would make the whole page inoperable/inaccessible, the new behavior doesn't.


Former-commit-id: 36a03295797b88190b87092e1a32d75ff4f45e12
Former-commit-id: 2b752600b10843da9f02230c9f891856bfafc7c4