## What is the error
'handleURLSlashes' get irregular result from a particular multibyte character.

## Symptoms

For some reason, when this multibyte character '移' is included in query string such as Search block, it doesn't get encoded properly.
Depends on the server, but when I tried to search '移' character, I get the following result

- re-direct loop
- no result
- or `htmlspecialchars(): Invalid multibyte sequence in argument` error.

When you URLencode '移', it support to become `%E7%A7%BB`
But for some reason, when handleURLSlashes() fires, it becomes `~%A7%BB`.

I'm sure there will be another multibyte characters that may cause the same error.

This '移' character means "move", so we use this character a lot.
So it's kinda critical that people cannot search using this character.

## First step

First, I just remove this function.

I just remove the handleURLSlashes for the quick fix.