Fix #3098
Fix #3098 by refactoring the Package class. Now all changes to the db.xml files (the package's own or its blocks') are respected if a package is re-installed (i.e. installed for a second time).

This also refactors the duplicate code in the Package class regarding installing the database through the db.xml.


Former-commit-id: f87fa6c412e52d50b7cd4d9ec9c298423748989f
Former-commit-id: b693f4cc7523079c9a463f4a06d7f5327b19c4aa