Fix #3098
Fix #3098 by refactoring the Package class. Now all changes to the db.xml files (the package's own or its blocks') are respected if a package is re-installed (i.e. installed for a second time).

This also refactors the duplicate code in the Package class regarding installing the database through the db.xml.