Feature: Attach mode for CLI installer
When the `--attach` flag is supplied with a `concrete5 c5:install` call, if the supplied database already has rows we will attach to it rather than failing.

This change adds an interface that lets you mark your own `StartingPointInstallRoutine` steps as required to run in attach mode.

This should resolve #3298