header_required cleanup
This will cleanup the header_required.php file. Changes:

- Code has been shortened;
- Code will be used consistent, instead of sometimes outputting straight JavaScript and in between echo'ing JavaScript in PHP;
- Variables have been removed, if they hadn't got to be in a variable but could be used straight away;
- Grouped meta tags into an array;
- Grouped shortcut links in an array;
- Description (meta) will only show when not empty (after trimming);
- Removed deprecated code (Loader::helper('concrete/ui'));

After this cleanup, we can think of firing an event or something, for Add-Ons to be able to add/edit some meta tags. For now, this cleanup only has parts it already had, but organized, clean, no deprecated code and will make it easier to add something in for Add-Ons at a later stage (commit/PR).