Multidimensional array - Idea for topic-list
You right about the white spacing. I will try to fix this.

I have "next generation" idea for this list. Instead of regular classes i create multidimensional array - with 7 classes (parent, nested, active <a>, active <li> + js attribute and so on - all that you need :) ).
Example (the code of each block is the output of the block from my C5 site):
http://codepen.io/siton/full/PzGpOP/

I create this 4 examples in 4-5 minutes (its really easy to change the classes like this to your "framework" markup). This examples could be "comment" in the core block. This is more "concept" issue, but for novice its easy to change the markup like this.

In cases like this materializecss or boostrap accordion markup (Extra divs and so on) anyway you need to create custom template:
http://materializecss.com/collapsible.html