Adjustment of thumbnail creation logic
Fixes the way thumbnails are generated when the source image has the same width as the thumbnail size.
With this change, a thumbnail is generated when:
- The source image is equal to or greater than the thumbnail width
- But if it's equal to the width and there's no thumbnail height or the thumbnail height is smaller than the image height, don't create the thumbnail

Previous to this change you could have images that were exactly matching the width of the thumbnail, but taller than thumbnail (i.e. it would be useful to apply cropping selection) not having a thumbnail generated.


Former-commit-id: 812dbd91bb92bc9f36c451ad58d034a275c57b7c
Former-commit-id: fe90bab15674767ea3c321a817145e669e070e6b