Fix #4658
What I noticed is when a task is used directly, as it is in the form
block, the bID is added automatically at the end. You can see it in the
action() function in the BlockView class.

Thanks to that, @mlocati modification worked like a charm.

The problem with the topic list is the page gets the information first
through its controller pageController.

In that class, through the function validateRequest() it first checks
the task by sending it to its own isValidControllerTask() function.

If it is not a valid task for the page, it then lists all the blocks and
one by one call their isValidControllerTask() function.

This is there the bID is missing and what makes @mlocati improvement
fail.

I modified that function to add the bID to the parameters array and it
now works with @mlocati original modification.