Fix #4658
@mlocati rightfully pointed out that when using a block's task we needed
a way to know which instance of the block launched the call.

He submitted a solution that worked in some situations but not all.

For instance, it broke the topic list block because the action was
called from the page's controller and missing the bID parameter.

It also didn't work when tasks were called from a block's Add (no bID)
or edit mode.

This pull request builds on top of his solution.

@mlocati solution only worked when the check on the task was done by the
block's controller. In that case the checking function automatically
adds the bID to the parameters and @mlocati check works.

But sometimes, the task will first go through the page's controller.

I modified the task checking function in the page's controller to also
add the bID to the list of parameters before running the check. If the
check is successful, the bID is removed from parameters and the task is
processed.
My modification also deal with a block's add and edit mode.