Fix localization issue with Punic
This fixes and issue where Punic's locale was set incorrectly
in case the active context was reverted back using the
`popActiveContext()` method. In this situation, the Punic's
locale was left to the previous locale set when
`pushActiveContext()` got called. This caused the Punic
translated site context's translations to be incorrect.