Update DateTime.php date function returns ''
Because the date $value is an instance of PHPDateTime the code was never setting $dateTime and therefore made the Javascript blank and the date picker would never load the value it had.  I made the functions DateTime and Date mimic each other in that respect.  

I don't not understand the reason for this code, but I left it in place.  Based on my understanding it will never be executed under any conditions.

} elseif ($value !== '') {
            $dateTime = new PHPDateTime('now', $dh->getTimezone('user'));    
        }