Lightbox in CKeditor link plugin
In 5.7, there was a possibility to make a link to a page or an image open in a lightbox using Magnific pop-up.
Since all the code was still there, I modified concrete5link plugin.js to get the lightbox back in the editor.

The only difference is I didn't make it an option to switch on or off. It's just included.