Update survey block controller.php
Surveys are not working properly in this version. When you vote on a survey you do not see any results and you can not view the results in the Dashboard. After digging through the code and database it appears that the cID (collection ID) is not being set properly when setPollOptions() is called. Added this to the action_form_save_vote() method and things appear to be working as expected now.