Avoid looking for EXIF metadata twice
Rotating an image using the EXIF metadata is done twice: in the ImageInspector
and in the AutorotateImageProcessor.

Both operations are performed when importing an image and when re-analyzing it,
so it's save to use just one of them.
Let's to that only in the processor (since rotating an image is a process, not
an inspection).