Update AttributeKeyView.php
If an AttributeValueObject doesn't exist for a given AttributeKey when being loaded into some Context then create a default ExpressValue. This allows the Attribute Controller to handle the empty value. Without the default ExpressValue the 'value' ScopeItem object is of the wrong type and the Attribute Controller is not loaded and can not handle the null case.