fix #9053 by not killing the whole page
By returning an empty FileFolder object, the page will still render, and the now missing folder will no longer be associated with the block. This conditional block is only called in three methods: `action_upload`, `formatBreadcrumbs` and `setupFolderFileFolderFilter`. The other methods call this method with `$realRoot = true`, thus bypassing this block. 

All three of these methods fail gracefully with this change. If the page is already loaded, then someone deletes the folder before another user uploads a file, the page reloads and the block shows no files found. `formatBreadcrmbs` and `setupFolderFileFolderFildter` similarly fail gracefully. 

It may be better to add a warning that a document library block depends on a folder when deleting it, but this at least doesn't cause the page to fail catastrophically.