Bugfix: make "show folders" option work without filesets
New "show folders" option only worked when a file set was also selected. However, it's not clear from the label that these two settings go together. It would be more self-explanatory if the "show folders" option worked independently of other settings. This change simply conditionally modifies the query in the view method based on `$this->hideFolders`.