Bugfix: anchor target attribute can be incorrect
The anchor target attribute will only be set correctly for the first page in the list because in the check to set it to "_self" (the default), `$target` has to be null. But it's only null on the first iteration. I changed this so that it's "_self" be default and then only changes if certain conditions are met.