Prevent ReCAPTCHA timeout
Issue 9780

When a form using reCAPTCHA V3 is loaded a time-out timer starts running. After 2 minutes the code expires and every form submission after this 2 minutes is not accepted and an error is given.

All we have to do is call grecaptcha.ready again a few seconds before it times out.