Avoid class::method error in marketplace update
#10014 describes a whoops arising from flawed code in the error handling for errors when connecting to the marketplace 
This change is untested because testing would require the initial error condition. It adds checks to avoid class::method error inside the exception handling for marketplace update.

As a fix, I don't know if the calling code to which the Error or ErrorList object is returned can handle an ErrorList (line 230) 
This is a messy fix. Better would be to restructure this method to avoid variant object type for $pkg. (line 243)