Allow for null responseObject
A site was throwing a method on null exception for $responseObject, presumably due to a dodgy file somewhere before this. 
Adding an is_object check to the condition prevents the exception. This does not cure the original problem, somewhere further back a non-file is getting passed in. I have not been able to trace that. However, it does stop the exception.